+++ Entity/Channel.php
@@ -389,6 +389,14 @@ class Channel extends ApiEntity implements GroupableEntity
         $slideOrders = $this->getChannelSlideOrders()->matching($criteria);
         foreach ($slideOrders as $slideOrder) {
             $slide = $slideOrder->getSlide();
+
+            if (is_null($slide)) {
+                $logger = $this->getContainer()->get('os2display.core.channel');
+                $logger->error('Os2Display\CoreBundle\Entity\Channel: Missing slide for channel, continuing: ' . $this->getId());
+
+                continue;
+            }
+
             if ($slide->isSlideActive()) {
                 $result->add($slide);
             }